- Added modoptionsapi.TextInputField to abstract text input guis - Added modoptionsapi.KeyBindingField for a key binding GUI (similar to Notch's) - Made many updates and bug fixes in order to get version to work - Tested fully
This version shouldn't be used. It is bugged in the sense that I used "Character" for a key input value instead of an Integer. This can essentially be rectified by find and replace, but I am committing the currently working code in order to facilitate any damages done by Find and Replace.